Move UV setup helpers into uv.go (match neighbor file pattern)
pip.go has CreatePipConfigManually, poetry.go has ConfigPoetryRepo —
setup helpers belong in the PM's own file, not a separate uvsetup.go.

- Merge uvsetup.go contents into uv.go
- Rename uvsetup_test.go → uv_test.go
